Added initial impulse and modified Player functio Update

+++ SDL2_OpenGL_BoilerPlate/BoilerPlate/MathUtilities.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-static float PI = 3.14159;
+static double PI = 3.14159;
 
 class MathUtilities{
 	public:
